backend/requirements.txt
# ============================================================================
# BACKEND DEPENDENCIES - Real-Time System Monitoring
# ============================================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Time-Series Database
influxdb-client==1.18.0

# Caching & Message Queue
redis==5.0.1
python-socketio==5.10.0
python-engineio==4.8.0

# API Integration
requests==2.31.0
aiohttp==3.9.1
httpx
websockets==12.0

# System monitoring
psutil==5.9.6

# ML & Data Processing
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2

# Deep Learning / Hugging Face
torch==2.2.0
transformers==4.36.0
datasets==2.14.5
sentence-transformers==2.2.2

# GROQ API Client
groq==0.4.2

# Ollama Integration
ollama==0.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
PyJWT==2.8.0

# API Documentation
python-multipart==0.0.6

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
typer==0.9.0
python-dateutil==2.8.2
pytz==2023.3
slugify==0.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Performance
orjson==3.9.11
ujson==5.8.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pylint==3.0.3

# Optional: Kafka support for large-scale deployments
kafka-python==2.0.2

# Optional: Message brokers
celery==5.3.4
kombu==5.3.4

# Optional: APM & Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-prometheus==0.42b0
